ADMIN 권한의 사용자는 게시글 및 댓글 수정과 삭제 가능 기능 구현

+++ src/main/java/com/example/post2/service/CommentService.java
@@ -2,7 +2,6 @@ package com.example.post2.service;
 
 import com.example.post2.dto.CommentRequestDto;
 import com.example.post2.dto.CommentResponseDto;
-import com.example.post2.dto.PostRequestDto;
 import com.example.post2.dto.StatusResponseDto;
 import com.example.post2.entity.Comment;
 import com.example.post2.entity.Post;
@@ -28,7 +27,7 @@ public class CommentService {
         this.commentRepository = commentRepository;
     }
 
-
+    // 댓글 작성
     public CommentResponseDto createComment(Long id, CommentRequestDto requestDto, UserDetailsImpl userDetails) {
         Post post = findPost(id);
 
@@ -41,8 +40,6 @@ public class CommentService {
         // 게시글의 댓글 리스트에 댓글 추가하기
         post.addComment(comment);
 
-        // 정렬한 내용을 할 때마다 넣어주어야 하는가 ..? 아니면 그냥 정렬하기만 하면 되는건가 ?
-
         CommentResponseDto commentResponseDto = new CommentResponseDto(comment);
         return commentResponseDto;
     }
@@ -50,11 +47,12 @@ public class CommentService {
     @Transactional
     public CommentResponseDto updateCommnet(Long id, Long commentid, CommentRequestDto requestDto, UserDetailsImpl userDetails) {
         // 해당 게시글이 있는지 확인
-        Post post = findPost(id);
+        findPost(id);
         // 해당 댓글이 있는지 확인
         Comment comment = findComment(commentid);
-        // 댓글 작성자와 Token의 작성자가 같은지 확인
-        if (comment.getUser().getUsername().equals(userDetails.getUsername())) {
+        // 해당 댓글을 작성한 작성자 이거나, 권한이 ADMIN인 경우 댓글 수정 가능
+        if (comment.getUser().getUsername().equals(userDetails.getUsername())
+                && userDetails.getUser().getRole().getAuthority().equals("ADMIN")) {
             // 있으면 댓글 내용 업데이트
             comment.update(requestDto);
             // ResponseDto에 내용 담아서 반환
@@ -67,12 +65,16 @@ public class CommentService {
     }
 
     public StatusResponseDto deleteComment(Long id, Long commentid, UserDetailsImpl userDetails) {
-        Post post = findPost(id);
+        // 해당 게시글이 있는지 확인
+        findPost(id);
+        // 해당 댓글이 있는지 확인
         Comment comment = findComment(commentid);
-
-        if (comment.getUser().getUsername().equals(userDetails.getUsername())) {
+        // 해당 댓글을 작성한 작성자 이거나, 권한이 ADMIN인 경우 댓글 삭제 가능
+        if (comment.getUser().getUsername().equals(userDetails.getUsername())
+                && userDetails.getUser().getRole().getAuthority().equals("ADMIN")) {
+            // 있으면 댓글 삭제
             commentRepository.delete(comment);
-
+            // 상태 ResponseDto에 담아서 반환
             StatusResponseDto statusResponseDto = new StatusResponseDto();
             statusResponseDto.setMsg("댓글 삭제 성공");
             statusResponseDto.setStatusCode(200);
@@ -82,13 +84,12 @@ public class CommentService {
         }
     }
 
-
     // 해당 포스트를 찾아서 반환
     private Post findPost(Long id) {
         return postRepository.findById(id).orElseThrow(() ->
                 new IllegalArgumentException("선택한 게시글은 존재하지 않습니다."));
     }
-
+    // 해당 댓글을 찾아서 반환
     private Comment findComment(Long commentid) {
         return commentRepository.findById(commentid).orElseThrow(() ->
                 new IllegalArgumentException("선택한 댓글은 존재하지 않습니다"));
